Clean up repl code a bit
