+++ interpreter/src/main.rs
@@ -13,9 +13,6 @@ use std::process;
 
 use clap::{App, Arg};
 
-use rustyline::error::ReadlineError;
-use rustyline::Editor;
-
 mod arena;
 mod array;
 mod bytecode;
@@ -45,7 +42,7 @@ mod vm;
 
 use crate::error::RuntimeError;
 use crate::memory::Memory;
-use crate::repl::RepMaker;
+use crate::repl::repl;
 
 /// Read a file into a String
 fn load_file(filename: &str) -> Result<String, io::Error> {
@@ -61,62 +58,6 @@ fn read_file(filename: &str) -> Result<String, RuntimeError> {
     let contents = load_file(&filename)?;
 
     Ok(contents)
-}
-
-/// Read a line at a time, printing the input back out
-fn read_print_loop() -> Result<(), RuntimeError> {
-    // establish a repl input history file path
-    let history_file = match dirs::home_dir() {
-        Some(mut path) => {
-            path.push(".evalrus_history");
-            Some(String::from(path.to_str().unwrap()))
-        }
-        None => None,
-    };
-
-    // () means no completion support (TODO)
-    // Another TODO - find a more suitable alternative to rustyline
-    let mut reader = Editor::<()>::new();
-
-    // Try to load the repl history file
-    if let Some(ref path) = history_file {
-        if let Err(err) = reader.load_history(&path) {
-            eprintln!("Could not read history: {}", err);
-        }
-    }
-
-    let mem = Memory::new();
-    let rep_maker = RepMaker {};
-    let rep = mem.mutate(&rep_maker, ())?;
-
-    // repl
-    loop {
-        let readline = reader.readline("> ");
-
-        match readline {
-            // valid input
-            Ok(line) => {
-                reader.add_history_entry(&line);
-                mem.mutate(&rep, line)?;
-            }
-
-            // some kind of program termination condition
-            Err(e) => {
-                if let Some(ref path) = history_file {
-                    reader.save_history(&path).unwrap_or_else(|err| {
-                        eprintln!("could not save input history in {}: {}", path, err);
-                    });
-                }
-
-                // EOF is fine
-                if let ReadlineError::Eof = e {
-                    return Ok(());
-                } else {
-                    return Err(RuntimeError::from(e));
-                }
-            }
-        }
-    }
 }
 
 fn main() {
@@ -139,7 +80,9 @@ fn main() {
         // TODO
     } else {
         // otherwise begin a repl
-        read_print_loop().unwrap_or_else(|err| {
+        let mem = Memory::new();
+        let result = mem.enter(repl);
+        result.unwrap_or_else(|err| {
             eprintln!("Terminated: {}", err);
             process::exit(1);
         });
